Change how template lookup works
Until now, template paths were concatenated to `base-dir`.  Which is
the root of the *source* of the static site itself.  This commit
changes the base path for a template to be a list of paths.  The first
item on this path is the directory `templates` within the root of the
blorg module.  That way, a static site instance can add a new
`templates` directory to itself and add that to the template path and
override and/or reuse files that blorg provide off the shelf.

Same explanation, with ascii.

blorg source package
- blorg.el
- templates/
  - archive.html
  - blog.html
  - category.html
  - layout.html
  - post.html

static site instance
- publish.el
- templates/
  - post.html
- posts/
  - welcome.org

Let's say that both `post.html` templates {% extends "layout.html" %}
in their first line.  With that situation in mind, if `publish.el` has
a `blog-cli` call that calls for `post.html`, it will use the
`post.html` file within the templates directory of the *static site
instance*.  And the `layout.html` template used by `post.html`, as
well as all the other templates, will be found within blorg's source
package.