Add :input-aggregate to blog-cli to aggregate input list
This feature allows creating, for example, category and archive
pages.  This PR ships with a default function that does not aggregate
anything and is used by default (blorg-input-aggregate-none).

It also ships with a function for aggregating blogs per category. The
categories are read from the FILETAGS Org-Mode attribute and category
pages can be created to display all blog posts of a given category.