fix: only asset discovery needs recursive search of files
Let's say one has the following file structure:

  a-blorg
  ├─ index.org
  ├─ other-stuff/
  │  └─ index.org
  └─ more-stuff/
     └─ index.org

And let's say one wants to generate an `index.html' that does not
include anything within either `other-stuff' or `more-stuff'.  The
route would have matched *all* `index.org' files before this change,
which could be a bit confusing:

 (blorg-route
  :name "index"
  :input-pattern "index.org"
  :template "index.html"
  :output "index.html"
  :url "/")

So, this PR changes `blorg--find-source-files' to take a new optional
parameter RECURSIVE which defaults to false, so the above code will do
for the case described above.

However, when copying static assets (within `blorg-export-assets'), we
most likely want to find all the files recursively within a directory
so we changed the call to `blorg--find-source-files' to pass `t' to
the RECURSIVE flag so it continues to work the way it currently does.