fix/feat: use globs for input-pattern
As the build of the last commit can tell, my solution for matching
specific files didn't work.  This new attempt fixes the tests and also
seems to handle all the cases I've seen so far, by matching:

 1. A specific file by just typing its relative path
 2. An expression that captures a collection of files
 2. Files within subdirectories, recursively or not

The best thing I know to solve all these problems is the good old GLOB
expressions, popular within shells.

Taking the exact example I used in my previous commit

  a-blorg
  ├─ index.org
  ├─ other-stuff/
  │  └─ index.org
  └─ more-stuff/
     └─ index.org

we can match many combinations of files with simple GLOBs:

  Expr          Files matched
  ----          -------------
  index.org     index.org
  */index.org   more-stuff/index.org, other-stuff/index.org
  **stuff/*     more-stuff/index.org, other-stuff/index.org
  **/index.org  index.org, more-stuff/index.org, other-stuff/index.org

The negative side is that this is a breaking change because if there
was regular expression character present in the `:input-pattern' of a
route before, it will have to be changed.  The good news is that I bet
there aren't many users of this tool already as we're in a very early
version of the blorg so it should be fine.

Notice though that `:input-exclude' still takes a regular expression
that will be used to filter out entries in the list returned by the
GLOB within `:input-pattern'.