refactoring: better handling for configuration errors
Basic use cases tackled:

  1. if `input-pattern` of a route doesn't match any files, this
  should be an error;

  2. if a route has an `input-pattern` that matches any files, but it
  doesn't have a `template`, it should be an error;

  3. if a route does not have an `input-pattern` it shouldn't bother
  trying to load any templates, thus `template` becomes optional;

The case N.3. above might be useful for adding routes that are managed
outside the blorg configuration.  One personal example: I have a
website with blog that are generated by blorg but there's one
subdirectory that I want to link to from within my website/blog but I
don't want blorg to generate the files within it.

This is the route I have there:

  (blorg-route
   :name "slides"
   :url "/slides")

And that's enough to link to that route with `{{ url_for("slides" )}}`
from any part of the website/blog.