refactoring: make blorg-export-templates plural, it's what it does
That got me confused for a little bit.  I started thinking that the
loop in where this function is called was looping through each
template instead of looping into each ROUTE.

This is still a better name than `blorg-route-export-templates`
because this function is public and it goes into a route
configuration, so adding the `route` sub-namespace would only make it
more of a mouthful than it already is.