feat: two new template filters: "default" & "len"
The `len' one is just a forward of the native #'length function. And
the `default' filter picks either the first value or the second if
the first one is nil.

Notice that all failures that happen within filters will most likely
explode without context where it happened.  Providing line template
and that a filter failure happens is an issue yet to be
solved (probably in templatel).