feat: enable automatic HTML entity escaping
If there's any variable to be rendered in a template with its content
not escaped, the template writer must mark such value as safe.  e.g.:

  <!-- template.html -->
  <div class="post">
    {{ post.html | safe }}
  </div>

All built-in filters are aware of safe values.  Custom filters must
be as well, so `safe` can fit in well with any other filter