feat: add optional anchor to links generated by url_for
The `url_for` function pulls the base URL of the website being
generated and concatenates to the `:url` field of the route encoded
in the URL.

In order to allow generating links with anchors, if the the keyword
argument `anchor` is not nil, it will be concatenated to the end of
`url_for` with a `#` in front of it.  e.g.:

  # for org mode files
  [[url_for:api,anchor=symbol-templatel-env-new][templatel-env-new]]

or

  <!-- for template files -->
  <a href="{{ url_for("api", anchor="symbol-templatel-env-new") }}">
    template-env-new
  </a>

Will both generate the following link:

  http://localhost:8000/api.html#symbol-templatel-env-new

That assumes a ~publish.el~ file with a route named "api" with an
":url" field with "api.html" that generates HTML with an actual anchor
link in that file.