feat: allow declaring variables visible to all templates in a site
If there is a set of variables that is used by, let's say,
`layout.html`, all the routes that use templates that inherit from
this base one will have to have those variables.

To make it less tedious for using weblorg, we're just adding the
`:template-vars` option to `weblorg-site` and then passing that into
every single template.