theme: autodoc should use single route "docs" for all things docs
The templatel documentation website is the guinea pig that we use to
drive the development of the autodoc theme.  On its lest iteration, I
found out that it'd be fairly easy to have just one "docs" route
instead of having both "doc" and "docs".  A simple if statement when
iterating over all the articles to make the side panel was enough to
get rid of the entire extra route.