feat: allow list of patterns for autodoc input-source
On top of taking a single string pattern to find symbols, this change
will allow users to pass lists of patterns.  The patterns on the list
will be resolved from left to right (with mapcan), so you can count on
the order. e.g.:

The following example will output the symbols lexicographically
sorted, which will return the `weblorg-copy-static' symbol first
and then the symbol `weblorg-route' later.

  (weblorg-input-source-autodoc
   "^weblorg-\\(?:route\\|copy-static\\)"

If order matters and one wants to render the documentation of
`weblorg-route' before the documentation of `weblorg-copy-static',
it's now possible to do the following:

  (weblorg-input-source-autodoc
   '("^weblorg-route" "^weblorg-copy-static"))