feat: make IDs anchor links of headlines off their text
This changes drastically the default behavior of Org-Mode to take the
headline text and use it to generate both IDs and anchors.  If
repeated IDs ever become an issue (when more than one headline has the
same text) then we can 1. just use CUSTOM_ID as usual and put whatever
they want; 2. maybe we add a map to keep the counts of a certain ID so
we can append its count to the final string; didn't do that yet to
keep complexity down until someone actually needs it;