refactor: change base directory for templates
This prepares us for a breaking change.  Before this change, we were
looking up templates first in the `<site-root>/templates` directory
and although it has its merits of being simple, it can be confusing
for people that have more experience building themes.

Another reason for this change is that we will need a place to put
static assets somewhere within the site-root directory preferrably not
sharing them with directory we export everything.  That will allow us
to override asset files from a theme we're inheriting from by just
having its relative path and name duplicated in the local assets
directory.

To prevent breaking stuff for people that have already created
weblorgs, we're marking this option as deprecate, the documentation
has been changed, and we'll do some more work to make sure it's easy
for people to figure out that it has changed.