[breaking] refactor/feat: PATH for theme's templates & static files
There has been a bit of cruft around the code that selects static
files to be copied from either external themes or site themes.  This
was improved by introducing the private function `weblorg--path`, that
is responsible for building the path from first the user's site and
then from the selected theme.

Another big and breaking change is that now the `:theme` parameter for
both `weblorg-site`, `weblorg-route` and `weblorg-copy-static` will
take a function instead of taking a string.  That function must
return the root directory of the theme one wants to use.  e.g.:

    (weblorg-site :theme #'weblorg-theme-default)