feat: add `post.url` property to parsed .org files
It is now possible to access the rendered `:url` property of each Org
file by using `{{ post.url }}` instead of the usual:

   `{{ url_for("route-name", param=value, ...) }}`

That's possible to be done automatically because we have all the
values that might be needed to fill up the `:url` property before
rendering the `:template` property.  That makes it more convenient for
theme authors to type less and get the same outcome.

Behind the scenes, it took adding a new helper function for rendering
a string within a template environment (weblorg--render-route-prop)
and also add a template environment for the (weblorg-copy-static)
routes.  This feature required the bumping of templatel to 0.1.5
because we needed `templatel-env-remove-template', which was just
added to allow this task to be completed.

Refs #29