feat: inject `OPTIONS: ^:{}` by default in autodoc
Before autodoc processes the doc string of a symbol, it will now
prepend a header to it.  As before, the doc string is interpreted as
an Org Mode document.  So, the header just injects a default option
that will disable the quick syntax for subscript/superscript so one
can use underscores and carets with no concern.  Formatting text as
subscript/superscript can still be done with _{} and ^{}.

Some escaping issues related to single quotes within doc strings were
also fixed.